requirements.txt
agora-token-builder==1.0.0
aiopg==1.4.0
amqp==5.2.0
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
autobahn==23.6.2
Automat==22.10.0
autopep8==2.0.4
beautifulsoup4==4.12.2
billiard==4.2.0
celery==5.3.6
certifi==2023.11.17
cffi==1.16.0
channels==3.0.5
channels-postgres==0.0.3
channels-redis==4.1.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
constantly==23.10.4
cryptography==41.0.7
daphne==3.0.2
Django==4.2.7
django-autoslug==1.9.9
django-bootstrap4==23.2
django-cors-headers==4.3.1
django-crispy-forms==2.1
django-filter==23.5
django-model-utils==4.3.1
django-notifications-hq==1.8.3
django-phonenumber-field==7.2.0
django-shortuuidfield==0.1.3
django-twilio==0.14.3.2
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
extensions==0.4
gunicorn==21.2.0
h11==0.14.0
hyperlink==21.0.0
idna==3.6
IMAPClient==3.0.1
incremental==22.10.0
jsonfield==3.1.0
kombu==5.3.4
msgpack==1.0.7
ndg-httpsclient==0.5.1
packaging==23.2
phonenumbers==8.13.26
Pillow==10.1.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pusher==3.3.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycodestyle==2.11.1
pycparser==2.21
PyJWT==2.8.0
PyNaCl==1.5.0
pyOpenSSL==23.3.0
python-dateutil==2.8.2
python-decouple==3.8
python-dotenv==1.0.1
pytz==2023.3.post1
redis==5.0.1
requests==2.31.0
service-identity==23.1.0
shortuuid==1.0.11
six==1.16.0
soupsieve==2.5
sqlparse==0.4.4
stripe==7.8.1
swapper==1.3.0
toml==0.10.2
tomli==2.0.1
twilio==7.17.0
Twisted==23.10.0
txaio==23.1.1
typing==3.7.4.3
typing_extensions==4.8.0
tzdata==2023.4
urllib3==2.1.0
uvicorn==0.24.0.post1
vine==5.1.0
wcwidth==0.2.13
zope.interface==6.1
